requirements.txt
Django==2.2
PostgreSQL==11.5
python==3.7.1
djangorestframework==3.10.3